Added Rematch and Opponent Leaving
For local and online games, added a rematch feature. For online games specifically, the clients and server communicate and UI appears when a player leaves or when a player wants a rematch.

My next steps for the project are art improvements, additions. I wish to add Music, SFX. The UI also needs to look nicer.

Most likely will add more to the start screen, including art, credits, etc.